mgr/nfs: reorg cluster class and common helper methods
