+++ src/pybind/mgr/nfs/module.py
@@ -7,7 +7,8 @@ from mgr_module import MgrModule
 from volumes.module import mgr_cmd_wrap
 import orchestrator
 
-from .export.nfs import NFSCluster, ExportMgr
+from .export.nfs import ExportMgr
+from .export.cluster import NFSCluster
 
 log = logging.getLogger(__name__)
